make `BotData` into a dataclass and move loading functions to separate module
