+++ samanthas_telegram_bot/bot_data_loader/__init__.py
@@ -1,0 +1,5 @@
+from samanthas_telegram_bot.bot_data_loader.bot_data_loader import BotDataLoader
+
+__all__ = [
+    "BotDataLoader",
+]
