+++ samanthas_telegram_bot/bot_data_loader/phrases.csv
